+++ fmi-import/src/main/kotlin/no/ntnu/ihb/fmi4j/importer/fmi2/ModelExchangeInstance.kt
@@ -22,11 +22,11 @@
  * THE SOFTWARE.
  */
 
-package no.ntnu.ihb.fmi4j.importer.fmi2.me
+package no.ntnu.ihb.fmi4j.importer.fmi2
 
 import no.ntnu.ihb.fmi4j.FmiStatus
-import no.ntnu.ihb.fmi4j.importer.fmi2.AbstractFmuInstance
 import no.ntnu.ihb.fmi4j.importer.fmi2.jni.EventInfo
+import no.ntnu.ihb.fmi4j.importer.fmi2.jni.ModelExchangeLibraryWrapper
 import no.ntnu.ihb.fmi4j.modeldescription.ModelExchangeModelDescription
 
 /**
