+++ fmi-native/build.gradle
@@ -10,7 +10,7 @@ if (!project.getProperties().get("skipBuildNative", false)) {
 
     library { CppLibrary lib ->
 
-        baseName = "fmi2_jni"
+        baseName = "fmi4j"
 
         linkage = [Linkage.SHARED]
